pyproject.toml
[tool.poetry]
name = "cloudwatch-log-group-enforcer"
version = "0.1.0"
description = ""
authors = ["Alex Chesters <10558792+AlexChesters@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "cloudwatch_log_group_enforcer"}]

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "^1.25.4"
pyyaml = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
